wallet-connect-test-bench: Fix typecheck (#71)
TypeScript cannot infer that `inputParameter` is an array-case of `SmartContractParameters`, which in turn prevents it from inferring that `schema` has the correct type.

Adding a type assertion to `inputParameter` fixes this and thus the need for suppressing the `schema` type check.

A direct dependency to `@concordium/browser-wallet-api-helpers` has been added to appease the linter. This can be removed again once the type is re-exported in `wallet-connectors`.